Massive code enhancement. No longer dependant on math lib.
I re-designed a lot of the code and made the loop be optional infinite loop.

Multi-balls support. You can now display a bunch of balls.

Removed the math library's pow function and took a traditional route, making it faster and the code not rely on the math library in-order to get the squared distance of x from center of circle.

Changed the logic of how it should draw the circle by squaring the radius. This leads to better Circle shapes.